requirements.txt
numpy==1.19.5
torch==1.8.0
matplotlib==3.3.4
pandas==1.1.5
PyYAML==5.4.1
scikit_learn==0.24.1
